requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
keyrings.alt==4.1.0
pyicloud==1.0.0
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.27.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
zipp==3.8.0
